Commit 1 20/09/2016
This is the first commit of this project using git SCM.
All code is for the .exe in the folder "Executables".
Currently, this code works, and am waiting for errors to
be reported from users. Upon finding an error, I will
create another branch off this one (master) and merge
it with this one when I have fixed the bug.