Fixed Averaging Calculation Problem
Program was assigning averages wrong for a subgeading that had both
subsubheading and reqs. For example, if the reqs were 7,8 and the
average for a subsubheading's reqs was 12, the average calculated
for the subheading was ((7 +8)/2 + 12)/2 = 9.5. The Calulation should
be (7 +8 +12)/3 = 9

The difference here is that a subsubheading is also counted as a
requirement for its parent subheading